refactor : 서버와 연동 가능하도록 구조 수정
