+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AntiPhishingApp">
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.AntiPhishingApp">
